feat(CreateProject): switch to tabbed window
