+++ src/components/Projects/CreateProject/Tabs/CreateFile.vue
@@ -2,7 +2,7 @@
 	<ToggleSheet
 		:value="value"
 		@input="$emit('input', $event)"
-		:dark="true"
+		:dark="dark"
 		:isToggleable="isToggleable"
 	>
 		<template #default="{ value }">
@@ -48,6 +48,10 @@ export default {
 		SelectedStatus,
 	},
 	props: {
+		dark: {
+			type: Boolean,
+			default: true,
+		},
 		file: {
 			type: Object,
 			required: true,
